+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Added a method for manually loading a string table as a dictionary to DialogueRunner
+
+### Changed
+
+- Made the debug display in InMemoryVariableStorage slightly tidier
+- Made changing the InMemoryVariableStorage update its debug display's layout components
+- Made InMemoryVariableStorage's contents enumerable in a foreach loop
+- Fixed a bug where the Dialog would pause when a blocking command handler immediately calls its onComplete and returns
+- Fixed a bug where parsing the `<<wait>>` command's parameter was locale-specific (i.e. certain European locales parse decimal numbers as "1,0"), which meant that behaviour would vary based on the end-user's configuration.
+
 ## [1.0.1] - 2020-01-08
 
 A bugfix release.