+++ CHANGELOG.md
@@ -8,6 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+### Changed
+
+### Removed
+
+## [1.0.2] - 2020-01-23
+
+Bug fixes and small quality-of-life improvements.
+
+### Added
+
 - Added a method for manually loading a string table as a dictionary to DialogueRunner
 - DialogueUI now allows skipping to the end of a line's delivery, by calling MarkLineComplete before the line has finished appearing.
 - Option buttons can now use TextMeshPro Text components, in addition to Unity UI Text components. (TextMeshPro for line display was already supported.)